requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
idna==2.8
importlib-metadata==1.3.0
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.6
pytest==5.3.2
pytest-base-url==1.4.1
pytest-html==2.0.1
pytest-metadata==1.8.0
pytest-selenium==1.17.0
pytest-variables==1.9.0
requests==2.22.0
selenium==3.141.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
